feat: specialize-into-bytes ff

+++ src/commands/impls/server.rs
@@ -111,11 +111,11 @@ pub fn split(inner: &RefCount<RedisClientInner>) -> Result<Vec<RedisClient>, Red
 pub async fn force_reconnection(inner: &RefCount<RedisClientInner>) -> Result<(), RedisError> {
   let (tx, rx) = oneshot_channel();
   let command = RouterCommand::Reconnect {
-    server:                               None,
-    force:                                true,
-    tx:                                   Some(tx),
+    server: None,
+    force: true,
+    tx: Some(tx),
     #[cfg(feature = "replicas")]
-    replica:                              false,
+    replica: false,
   };
   interfaces::send_to_router(inner, command)?;
 
@@ -150,7 +150,7 @@ pub async fn ping<C: ClientLike>(client: &C) -> Result<RedisValue, RedisError> {
 }
 
 pub async fn select<C: ClientLike>(client: &C, db: u8) -> Result<RedisValue, RedisError> {
-  let frame = utils::request_response(client, || Ok((RedisCommandKind::Select, vec![db.into()]))).await?;
+  let frame = utils::request_response(client, || Ok((RedisCommandKind::Select, vec![(db as i64).into()]))).await?;
   protocol_utils::frame_to_results(frame)
 }
